Rename _transcoding module, and deprecate constant
