+++ tests/pipeline/test_pipeline.py
@@ -5,13 +5,13 @@ import pytest
 
 import kedro
 from kedro.pipeline import node
-from kedro.pipeline._transcoding import _strip_transcoding, _transcode_split
 from kedro.pipeline.modular_pipeline import pipeline as modular_pipeline
 from kedro.pipeline.pipeline import (
     CircularDependencyError,
     ConfirmNotUniqueError,
     OutputNotUniqueError,
 )
+from kedro.pipeline.transcoding import _strip_transcoding, _transcode_split
 
 
 class TestTranscodeHelpers:
