+++ tests/pipeline/test_pipeline_with_transcoding.py
@@ -4,12 +4,12 @@ import pytest
 
 import kedro
 from kedro.pipeline import node
-from kedro.pipeline._transcoding import _strip_transcoding
 from kedro.pipeline.modular_pipeline import pipeline as modular_pipeline
 from kedro.pipeline.pipeline import (
     CircularDependencyError,
     OutputNotUniqueError,
 )
+from kedro.pipeline.transcoding import _strip_transcoding
 
 
 # Different dummy func based on the number of arguments
